+++ CourseSchedulingSystem/Pages/Manage/ScheduleTypes/Delete.cshtml.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using CourseSchedulingSystem.Data;
 using CourseSchedulingSystem.Data.Models;
@@ -19,11 +20,17 @@ namespace CourseSchedulingSystem.Pages.Manage.ScheduleTypes
 
         [BindProperty] public ScheduleType ScheduleType { get; set; }
 
+        public bool InUse => ScheduleType.CourseScheduleTypes.Any();
+
         public async Task<IActionResult> OnGetAsync(Guid? id)
         {
             if (id == null) return NotFound();
 
-            ScheduleType = await _context.ScheduleTypes.FirstOrDefaultAsync(m => m.Id == id);
+            ScheduleType = await _context.ScheduleTypes
+                .Include(st => st.CourseScheduleTypes)
+                .ThenInclude(cst => cst.Course)
+                .ThenInclude(c => c.Subject)
+                .FirstOrDefaultAsync(m => m.Id == id);
 
             if (ScheduleType == null) return NotFound();
             return Page();
@@ -33,10 +40,17 @@ namespace CourseSchedulingSystem.Pages.Manage.ScheduleTypes
         {
             if (id == null) return NotFound();
 
-            ScheduleType = await _context.ScheduleTypes.FindAsync(id);
+            ScheduleType = await _context.ScheduleTypes
+                .Include(st => st.CourseScheduleTypes)
+                .FirstOrDefaultAsync(m => m.Id == id);
 
             if (ScheduleType != null)
             {
+                if (InUse)
+                {
+                    return Page();
+                }
+
                 _context.ScheduleTypes.Remove(ScheduleType);
                 await _context.SaveChangesAsync();
             }
