+++ CourseSchedulingSystem/Pages/Manage/Users/Delete.cshtml.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using CourseSchedulingSystem.Data;
 using CourseSchedulingSystem.Data.Models;
@@ -20,6 +21,8 @@ namespace CourseSchedulingSystem.Pages.Manage.Users
         [BindProperty]
         public User User { get; set; }
 
+        public bool CanDelete { get; set; }
+
         public async Task<IActionResult> OnGetAsync(Guid? id)
         {
             if (id == null)
@@ -36,6 +39,9 @@ namespace CourseSchedulingSystem.Pages.Manage.Users
             {
                 return NotFound();
             }
+
+            CanDelete = await MoreThanOneActiveUser();
+            
             return Page();
         }
 
@@ -50,11 +56,23 @@ namespace CourseSchedulingSystem.Pages.Manage.Users
 
             if (User != null)
             {
+                if (!await MoreThanOneActiveUser())
+                {
+                    return Page();
+                }
+
                 _context.Users.Remove(User);
                 await _context.SaveChangesAsync();
             }
 
             return RedirectToPage("./Index");
+        }
+
+        private async Task<bool> MoreThanOneActiveUser()
+        {
+            return await _context.Users
+                .Where(u => !u.IsLockedOut)
+                .CountAsync() > 1;
         }
     }
 }
